+++ pkg/nanolib/synapse/src/main.ts
@@ -2,4 +2,4 @@ export * from './bootstrap.js';
 export * from './auto-destruct.js';
 export * from './directive-decorator.js';
 export * from './directive-class.js';
-export * from './query-decorator.js';
+export * from './tools-decorator.js';
