+++ pkg/nanolib/synapse/src/tools-decorator.ts
